Remove HList (superseded by List)
